+++ apps/web/app/(user)/accounts/[id]/page.tsx
