+++ apps/web/app/(user)/layout.module.css
@@ -1,0 +1,30 @@
+.wrapper {
+  display: flex;
+  min-height: calc(100vh - 48px - 32px);
+}
+
+.nav {
+  display: flex;
+  flex-direction: column;
+  width: 20%;
+  max-width: 240px;
+  gap: 8px;
+  padding-right: 16px;
+  margin-right: 16px;
+  border-right: 1px solid var(--color-border);
+}
+
+.content {
+  flex: 1;
+}
+
+.active::before {
+  content: '';
+  display: block;
+  position: absolute;
+  width: 4px;
+  background-color: var(--color-focus);
+  height: 24px;
+  margin-left: -16px;
+  border-radius: 2px;
+}
