+++ apps/web/app/(user)/providers/page.tsx
@@ -1,0 +1,82 @@
+/* eslint-disable @next/next/no-img-element */
+import { getUser } from '@/lib/getUser';
+import { db } from '@/lib/db';
+import { notFound, redirect } from 'next/navigation';
+import { cache } from 'react';
+import { Headline } from '@gw2treasures/ui/components/Headline/Headline';
+import { Table } from '@gw2treasures/ui/components/Table/Table';
+import { AuthorizationType } from '@gw2me/database';
+
+const getUserData = cache(async () => {
+  const session = await getUser();
+
+  if(!session) {
+    redirect('/login');
+  }
+
+  const user = await db.user.findUnique({
+    where: { id: session.id },
+    include: { sessions: true, providers: true },
+  });
+
+  if(!user) {
+    notFound();
+  }
+
+  return {
+    sessionId: session.sessionId,
+    user,
+  };
+});
+
+export default async function ProfilePage() {
+  const { sessionId, user } = await getUserData();
+
+  return (
+    <>
+      <Headline id="providers">Login Providers</Headline>
+      <Table>
+        <thead>
+          <tr>
+            <th>Provider</th>
+            <th>User</th>
+            <th>Created</th>
+          </tr>
+        </thead>
+        <tbody>
+          {user.providers.map((provider) => (
+            <tr key={`${provider.provider}-${provider.providerAccountId}`}>
+              <td>{provider.provider}</td>
+              <td>{provider.displayName}</td>
+              <td>{provider.createdAt.toISOString()}</td>
+            </tr>
+          ))}
+        </tbody>
+      </Table>
+
+      <Headline id="sessions">Sessions</Headline>
+      <Table>
+        <thead>
+          <tr>
+            <th>Session</th>
+            <th>Started</th>
+            <th>Last Active</th>
+          </tr>
+        </thead>
+        <tbody>
+          {user.sessions.map((session) => (
+            <tr key={session.id}>
+              <td>{session.info}{session.id === sessionId && ' (Current Session)'}</td>
+              <td>{session.createdAt.toISOString()}</td>
+              <td>{session.lastUsed.toISOString()}</td>
+            </tr>
+          ))}
+        </tbody>
+      </Table>
+    </>
+  );
+}
+
+export const metadata = {
+  title: 'Security'
+};
